requirement.txt
anyio==3.5.0
argon2-cffi==20.1.0
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==4.1.0
Bottleneck==1.3.5
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.0
click==8.0.3
colorama==0.4.6
cryptography==41.0.5
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
email-validator==2.0.0.post2
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.2
Flask==2.2.5
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
flit_core==3.6.0
fonttools==4.38.0
greenlet==3.0.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.2.0
install==1.3.5
ipykernel==6.15.2
ipython==7.31.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.11.2
jupyter-server==1.23.4
jupyterlab-pygments==0.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mlxtend==0.23.0
munkres==1.1.4
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.21.6
openpyxl==3.0.10
packaging==22.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.8.0
pymongo==4.5.0
pyparsing==3.1.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
pywin32==305.1
pywinpty==2.0.10
pyzmq==23.2.0
regex==2022.7.9
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.7.1
urllib3==2.0.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
wheel==0.38.4
wincertstore==0.2
WTForms==3.0.1
xlrd==2.0.1
zipp==3.15.0
